feat(frontend): add GraphQL API query page
This change adds a new page that allows dynamically querying the backend
(as the currently authenticated user) with GraphQL queries and viewing
the results.

This replaces the need for the backend to serve the GraphiQL interface,
and instead embeds GraphiQL in the frontend module (where it belongs).

NOTE: the API page will not work with this commit alone, and is
dependent on an upcoming backend commit.