feat: refactor graph
This change refactors the Graph hierarchy, such that the root queries
only provide the tenant(s), and the tenants provide further
tenant-scoped operations such as fetching of accounts, etc.

This change affects how the frontend communicates with the backend:
- it no longer needs to signal which tenant the user is logged into
  (it's part of the query)
- backend authorization determines query/mutation success, based on
  tenant-scoped roles of the token's user

This change also affects mutations by requiring a tenant ID argument for
most operations.