fix(deploy): fix (again) deployment status detection by Skaffold
Skaffold seems to have a bug where it gets stuck during detection of
current state of deployment. It seems to be related to how it parses
events from the cluster.

By deleting past events - esp. those with errors - it ensures Skaffold
properly detects the current state rather than old errors.

This fix used to be present inthe skaffold.yaml but was removed to test
if the Skaffold bug was fixed.

It was NOT.