chore(ci): Merge branch & PR workflows
This change merges the workflows that were triggered by branch pushes
(to build the Docker images and run unit tests) into the PR deployment
workflow.

This ensures that we have a single point to inspect results of changes,
and also better aligns our workflows with Dependabot, which does not
trigger branch events.

When the end-to-end testing job reports the results, it will report them
as PR comments when running in the context of a PR, and as a commit
comment when triggered by a "push" event.