fix(backend): Fix backend CORS support
The backend CORS support failed when client declared that it will send
the "X-Request-ID" header - and this header is indeed specified in the
generated OpenAPI specification, and thus fails the API playground as
well as actual requests that would have sent this header.

This change fixes that by adding the "X-Request-ID" header to the list
of accepted headers for CORS requests, and also removes the deprecated
"X-GreenSTAR-Tenant-ID" header which is no longer used.