fix(backend): Init job does not generate sample data
This is caused by incorrect assumption for the default value of the
(now deleted) "GENERATE_SAMPLE_DATA" flag; boolean values ALWAYS have
a default value of "false" (because in the CLI they are just specified
as "--my-flag" without any value...).

To fix the issue, this change modifies the flag to be in the negative -
by renaming it from "--generate-sample-data" to
"--disable-sample-data-generation".