refactor(build): Support "extraEnv" composition from multiple files
This change modifies the "extraEnv" Helm value of the different
workloads from an array of name/value to an object, so that it's
possible to override individual environment variables by using multiple
Helm values files.

It also disables OTEL SDK in CI environment from the
"chart-ci-values.yaml" file instead of the CI workflow.