fix(build): Split build/test workflows from result publishing workflows
This change completes the split between building & running PR artifacts
and performing validation - and the publishing of the results of the
tests into separate workflows.

This split is necessary in order to ensure there is no leakage of access
or secrets, since building & testing untrusted code is dangerous by
nature. Thus, PR workflows are split to two parts:

1. Build & test workflows: these are run with no permissions or access
   to secrets, but do publish their results as artifacts.
2. Results publishing workflows: these need more permissions by nature
   (e.g. to upload test reports to a cloud bucket) and thus are run with
   more permissions - but do not execute any untrusted code. These
   workflows are triggered by the completion (whether successful or not)
   of the build & test workflows.