fix(ci): update execution conditions for publish workflow
Adjust condition checks in the GitHub Actions `publish` workflow to ensure
it only executes if the triggering workflow `verify` was executed in response
to a pull request.

The reason this is needed is because this workflow posts a PR comment, and
thus must not be executed if the underlying workflow (`verify`) was run in
response to a `push` event (or anything else).