feat(scripts): support downloading cluster-state from specific workflow runs (#408)
This change improves the `download-cluster-state.sh` (and the 
`download-cluster-state` task) to support downloading cluster state from a
specific GitHub Action run, in addition to  existing support for the latest
run of the current branch.

This enables running the following:

```shell
$ task download-cluster-state -- 123
```

This will download the cluster-state artifact of the GitHub Actions workflow
run `123` instead of the latest run of the current branch.

To accommodate this change, the script now also creates a unique directory
named after the the workflow run ID (whether inferred from current branch of
provided explicitly).

Signed-off-by: Arik Kfir <arik@kfirs.com>